Bundle missing packages, rename lilbee.mcp, add automated QA harness
Rename src/lilbee/mcp.py -> mcp_server.py for the same reason as platform
-> system: the bare `mcp` name shadowed the third-party mcp/ package and
Nuitka's --include-package=mcp resolved to the lilbee submodule (a single
file) instead of the standalone package, fatally bailing with "'mcp' is a
module not a package".

Add explicit Nuitka includes that the binary needs at runtime but Nuitka's
import-following missed:
- mcp + mcp.* (FastMCP runtime imports use string dispatch in fastmcp.tools)
- litestar + litestar.* (logging dictConfig references
  litestar.logging.standard.LoggingQueueListener by string)

Add a MYPYC_FLAGS shell loop that picks up *__mypyc*.so helpers at
site-packages root. Nuitka's standard config has implicit-imports for
charset_normalizer's mypyc helper but not chardet 7.x's; globbing makes
this forward-compatible without one-off package configs.

Add tools/qa-binary.sh: 19-check automated QA matrix exercising CLI
(version + warm cache + help + self-check-extras + status + add + search),
HTTP (server + health + status + search + SIGTERM), MCP (JSON-RPC
initialize), TUI (chrome + Ctrl-Q quit), and cache (per-version dir
present after first run). Validates the binary end-to-end before each
release.
